pyproject.toml
[project]
name = "de-template"
version = "0.1.0"
description = "Data engineering lakehouse pipeline template (Flink + Iceberg + dbt)"
readme = "README.md"
requires-python = ">=3.11"

# Host-side runtime deps — config validation and pipeline scripts.
# Docker containers manage their own runtime deps independently.
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",      # dataset manifest (scaffold_dataset.py)
    "jinja2>=3.1",      # Jinja2 templates for dataset scaffolding
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-timeout>=2.0",
    "pytest-sugar>=0.9",
    "ruff>=0.4",          # lint + format
    "pyright>=1.1",       # type checking
    "pyarrow>=14.0",      # CI fixture generator (tests/fixtures/make_parquet.py)
]

# =============================================================================
# pytest configuration (replaces pytest.ini)
# =============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

# Makes `from config.settings import ...` and `from scripts.health import ...`
# importable without installing the project as a package.
pythonpath = ["."]

markers = [
    "integration: full pipeline tests requiring Docker (deselect with -m 'not integration')",
    "slow: tests that take >60s (usually E2E Docker tests)",
]

# Show short tracebacks by default; use --tb=long for debugging
addopts = "--tb=short"

# Default timeout per test (seconds).
# Integration tests can override per-class with @pytest.mark.timeout(N).
timeout = 60

# =============================================================================
# ruff — lint + format
# =============================================================================
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
# E/W: pycodestyle  F: pyflakes  I: isort  UP: pyupgrade
select = ["E", "F", "I", "UP", "W"]
ignore = [
    "E501",   # line too long — handled by formatter, not lint
]

[tool.ruff.lint.isort]
known-first-party = ["config", "scripts"]

# =============================================================================
# pyright — type checking
# =============================================================================
[tool.pyright]
pythonVersion = "3.11"
venvPath = "."
venv = ".venv"
include = ["config", "scripts", "tests"]
reportMissingImports = true
reportMissingTypeStubs = false   # third-party stubs often missing — not an error

# =============================================================================
# uv configuration
# =============================================================================
[tool.uv]
# Keep .venv inside the project root for IDE discoverability
# (most editors auto-detect .venv at the project root)
package = false   # this is a scripts-only project, not a distributable package
